vlan: work around crash during network teardown
When tearing down all network devices on a system with
multiple vlans and multiple laggs, we hit a case where
a vlan interface either has its trunk killed, or has no
vlan at all (it's not entirely clear which, from the
crash data), at the time the interface's transmit
routine is called to send a packet.

(In this particular case, syslog-ng is attempting to send
a UDP packet to another system via the vlan.)

The code is already prepared to discard the packet if
the parent attached to the trunk is not up and running,
so as a band-aid to keep from crashing, if we're missing
an ifv or trunk, treat this the same way as "parent down"
(discarding the packet and returning ENETDOWN).

Ticket: #17542
(cherry picked from commit df0e24ec4b7094454ae37092589e1fe5efe10422)